requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
asn1crypto==1.4.0
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
billiard==3.6.3.0
boto3==1.14.44
botocore==1.17.44
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cmake==3.18.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==3.3.1
Cython==0.29.21
Django==3.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-email-verification==0.0.7
django-extensions==3.1.3
django-filter==2.3.0
django-pandas==0.6.4
django-storages==1.11.1
djangorestframework==3.11.1
djangorestframework-files==1.1.0
djangorestframework-simplejwt==4.5.0
dlib==19.21.0
dnspython==1.16.0
docutils==0.15.2
drf-yasg==1.17.1
eventlet==0.27.0
extras==1.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
flower==0.9.5
gevent==20.6.2
google-api-core==1.22.2
google-auth==1.21.3
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
greenlet==0.4.16
#grpcio==1.32.0
gunicorn==20.0.4
html5lib==1.1
humanize==2.6.0
idna==2.10
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
kombu==4.6.11
lazy-object-proxy==1.4.3
lxml==4.6.1
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
mysqlclient==2.0.3
Naked==0.1.31
numpy==1.19.1
openapi-codec==1.3.2
packaging==20.4
pandas==1.2.2
pdfminer==20191125
pep8==1.7.1
Pillow==8.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.49
requests==2.24.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
shellescape==3.8.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
supervisor==4.2.1
toml==0.10.1
tornado==6.0.4
twilio==6.53.0
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.4
zope.event==4.4
zope.interface==5.1.0
facebook-sdk==3.1.0
